pyproject.toml
[tool.poetry]
name = "santa-bot"
version = "0.1.0"
description = ""
authors = ["arsenyvolodko <arvolodko@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.1"
sqlalchemy = "^2.0.36"
aiogram = "3.5.0"
psycopg2-binary = "^2.9.10"
asyncpg = "^0.30.0"
greenlet = "^3.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
